fix on tests to use PalletBalances force_set_balance instead of mint
