+++ pallets/template/src/tests.rs
@@ -11,13 +11,14 @@
 // This flag tells rust to only run this file when running `cargo test`.
 #![cfg(test)]
 
+use crate::mock::{new_test_ext, RuntimeEvent, RuntimeOrigin, TestRuntime};
 use crate::*;
-use crate::{self as pallet_kitties};
-use frame::deps::frame_support::runtime;
-use frame::deps::sp_io;
-use frame::runtime::prelude::*;
+// use frame::deps::frame_support::runtime;
+// use frame::deps::sp_io;
+// use frame::runtime::prelude::*;
 use frame::testing_prelude::*;
-use frame::traits::fungible::*;
+use mock::{PalletBalances, PalletKitties, System};
+// use frame::traits::fungible::*;
 
 type Balance = u64;
 type Block = frame_system::mocking::MockBlock<TestRuntime>;
@@ -34,71 +35,6 @@ const DEFAULT_KITTY: Kitty<TestRuntime> = Kitty {
     owner: 0,
     price: None,
 };
-
-#[runtime]
-mod runtime {
-    #[runtime::derive(
-        RuntimeCall,
-        RuntimeEvent,
-        RuntimeError,
-        RuntimeOrigin,
-        RuntimeTask,
-        RuntimeFreezeReason,
-        RuntimeHoldReason
-    )]
-    #[runtime::runtime]
-    /// The "test runtime" that represents the state transition function for our blockchain.
-    ///
-    /// The runtime is composed of individual modules called "pallets", which you find see below.
-    /// Each pallet has its own logic and storage, all of which can be combined together.
-    pub struct TestRuntime;
-
-    /// System: Mandatory system pallet that should always be included in a FRAME runtime.
-    #[runtime::pallet_index(0)]
-    pub type System = frame_system::Pallet<Runtime>;
-
-    /// PalletBalances: Manages your blockchain's native currency. (i.e. DOT on Polkadot)
-    #[runtime::pallet_index(1)]
-    pub type PalletBalances = pallet_balances::Pallet<Runtime>;
-
-    /// PalletKitties: The pallet you are building in this tutorial!
-    #[runtime::pallet_index(2)]
-    pub type PalletKitties = pallet_kitties::Pallet<Runtime>;
-}
-
-// Normally `System` would have many more configurations, but you can see that we use some macro
-// magic to automatically configure most of the pallet for a "default test configuration".
-#[derive_impl(frame_system::config_preludes::TestDefaultConfig)]
-impl frame_system::Config for TestRuntime {
-    type Block = Block;
-    type AccountData = pallet_balances::AccountData<Balance>;
-}
-
-// Normally `pallet_balances` would have many more configurations, but you can see that we use some
-// macro magic to automatically configure most of the pallet for a "default test configuration".
-#[derive_impl(pallet_balances::config_preludes::TestDefaultConfig)]
-impl pallet_balances::Config for TestRuntime {
-    type AccountStore = System;
-    type Balance = Balance;
-}
-
-// This is the configuration of our Pallet! If you make changes to the pallet's `trait Config`, you
-// will also need to update this configuration to represent that.
-impl pallet_kitties::Config for TestRuntime {
-    type RuntimeEvent = RuntimeEvent;
-    type WeightInfo = (); // TODO: investigate why this works
-}
-
-// We need to run most of our tests using this function: `new_test_ext().execute_with(|| { ... });`
-// It simulates the blockchain database backend for our tests.
-// If you forget to include this and try to access your Pallet storage, you will get an error like:
-// "`get_version_1` called outside of an Externalities-provided environment."
-pub fn new_test_ext() -> sp_io::TestExternalities {
-    frame_system::GenesisConfig::<TestRuntime>::default()
-        .build_storage()
-        .unwrap()
-        .into()
-}
 
 #[test]
 fn starting_template_is_sane() {
@@ -121,8 +57,18 @@ fn system_and_balances_work() {
         // We often need to set `System` to block 1 so that we can see events.
         System::set_block_number(1);
         // We often need to add some balance to a user to test features which needs tokens.
-        assert_ok!(PalletBalances::mint_into(&ALICE, 100));
-        assert_ok!(PalletBalances::mint_into(&BOB, 100));
+        assert_ok!(PalletBalances::force_set_balance(
+            RuntimeOrigin::root(),
+            ALICE,
+            100
+        ));
+        assert_ok!(PalletBalances::force_set_balance(
+            RuntimeOrigin::root(),
+            BOB,
+            100
+        ));
+        // assert_ok!(PalletBalances::mint_into(&ALICE, 100));
+        // assert_ok!(PalletBalances::mint_into(&BOB, 100));
     });
 }
 
@@ -438,7 +384,13 @@ fn do_buy_kitty_emits_event() {
             kitty_id,
             Some(1337)
         ));
-        assert_ok!(PalletBalances::mint_into(&BOB, 100_000));
+        // assert_ok!(PalletBalances::mint_into(&BOB, 100_000));
+
+        assert_ok!(PalletBalances::force_set_balance(
+            RuntimeOrigin::root(),
+            BOB,
+            100_000
+        ));
         assert_ok!(PalletKitties::buy_kitty(
             RuntimeOrigin::signed(BOB),
             kitty_id,
@@ -490,13 +442,24 @@ fn do_buy_kitty_logic_works() {
             frame::arithmetic::ArithmeticError::Underflow
         );
         // Cannot buy kitty if it would kill your account (i.e. set your balance to 0).
-        assert_ok!(PalletBalances::mint_into(&BOB, 1337));
+        // assert_ok!(PalletBalances::mint_into(&BOB, 1337));
+
+        assert_ok!(PalletBalances::force_set_balance(
+            RuntimeOrigin::root(),
+            BOB,
+            1337
+        ));
         assert!(
             PalletKitties::buy_kitty(RuntimeOrigin::signed(BOB), kitty_id, 1337).is_err(),
             // TODO: assert_noop on DispatchError::Token(TokenError::NotExpendable)
         );
         // When everything is right, it works.
-        assert_ok!(PalletBalances::mint_into(&BOB, 100_000));
+        // assert_ok!(PalletBalances::mint_into(&BOB, 100_000));
+        assert_ok!(PalletBalances::force_set_balance(
+            RuntimeOrigin::root(),
+            BOB,
+            101_337
+        ));
         assert_ok!(PalletKitties::buy_kitty(
             RuntimeOrigin::signed(BOB),
             kitty_id,
@@ -509,7 +472,7 @@ fn do_buy_kitty_logic_works() {
         // Price is reset to `None`.
         assert_eq!(kitty.price, None);
         // BOB transferred funds to ALICE.
-        assert_eq!(PalletBalances::balance(&ALICE), 1337);
-        assert_eq!(PalletBalances::balance(&BOB), 100_000);
+        assert_eq!(PalletBalances::free_balance(&ALICE), 1337);
+        assert_eq!(PalletBalances::free_balance(&BOB), 100_000);
     })
 }
